do initial sketches of default layout & styling.
A CSS stylesheet and a "default" layout have been sketched for this
theme.  For now all it should do is center the content in a single 60%
width column, but I've also done some styling for the header element I
want to implement at the top of the page.

Content headers have also been formatted to be centered and have various
styles of underlines.